Final Fantasy Tactics style map (incomplete)
A map with hard coded node positions and weights. (Prefab:
MapFixedNodes) Node neighbours are not hard coded so nodes will use
their neighbours within a set radius but once we have a final map layout
it will probably be better to hard code it. Also renamed some prefabs
and scripts (sorry). I built in some error checking but it is still
prone to incomplete maps and intersecting paths.